pyproject.toml
[tool.poetry]
name = "server-bot"
version = "1.0.0"
description = "A private bot meant for Sham's Code Workspace."
authors = ["Sham <thesham@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.4.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
